Make startColumn option work in parseTable
Specifying a non-zero startColumn did not work correctly in parseTable (it would never pass the innermost conditional check and so never get incremented, resulting in an empty columns array). It now follows the format used in parseCSV. A call to tr.childNodes has been replaced by a call to tr.children as tr.childNodes includes whitespace "nodes" that mess up the function.